+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog for package grepros
 0.2.6 (2021-12-04)
 -------------------
 * add --write-format postgres
+* add --no-verbose option
 * use up to nanosecond precision in HTML output timeline
 * fix potential error on using --max-per-topic with live topics